Create Colorp_percept_circadianType
Create a new marix, numericalDATA, of the same size as DATA. It should represent the original data, but as numbers. Go through all data, parse it and assign numerical labels that you store in the numericalDATA matrix. Makes a figure that shows a histogram of the self-reported color responses.
Makes a figure that shows a histogram of the self-reported circadian type.